refactoring __init__
